wip mod ether net
