+++ src/lib.rs
@@ -103,109 +103,8 @@ pub mod bitfield;
 pub mod eth;
 pub mod icmp;
 pub mod ip;
-pub mod qinq;
 pub mod tcp;
 pub mod udp;
-pub mod vlan;
 pub mod vxlan;
 pub mod ne;
 pub mod types;
-
-/// Protocol which is encapsulated in the payload of the Ethernet frame.
-///
-/// According [EtherType](https://en.wikipedia.org/wiki/EtherType)
-#[repr(u16)]
-#[derive(PartialEq, Eq, Debug, Copy, Clone)]
-#[cfg_attr(feature = "serde", derive(::serde::Serialize, ::serde::Deserialize))]
-pub enum EtherType {
-    Loop = 0x0060,
-    Ipv4 = 0x0800,
-    Arp = 0x0806,
-    /// wake on lan
-    WakeOnLan = 0x0842,
-    /// Cisco Discovery Protocol
-    CDP = 0x2000,
-    /// Stream Reservation Protocol
-    SRP = 0x22EA,
-    /// Audio Video Transport Protocol
-    AVTP = 0x22F0,
-    /// IETF TRILL Protocol
-    TRILL = 0x22F3,
-    /// DEC MOP RC
-    MOP = 0x6002,
-    /// DECnet Phase IV, DNA Routing
-    DECnet = 0x6003,
-    DECLAT = 0x6004,
-    /// Reverse Address Resolution Protocol
-    RARP = 0x8035,
-    AppleTalk = 0x809B,
-    /// AppleTalk Address Resolution Protocol
-    AARP = 0x80F3,
-    /// VLAN-tagged frame (IEEE 802.1Q) and Shortest Path Bridging IEEE 802.1aq with NNI compatibility
-    VLAN = 0x8100,
-    /// Simple Loop Prevention Protocol
-    SLPP = 0x8102,
-    /// Virtual Link Aggregation Control Protocol
-    VLACP = 0x8103,
-    /// Internet Protocol Version 6
-    Ipv6 = 0x86DD,
-    MPLSUnicast = 0x8847,
-    MPLSMulticast = 0x8848,
-    /// Ethernet Slow Protocols such as the Link Aggregation Control Protocol (LACP)
-    LACP = 0x8809,
-    /// Service VLAN tag identifier (S-Tag) on Q-in-Q tunnel
-    QinQ = 0x88A8,
-    /// Link Layer Discovery Protocol
-    LLDP = 0x88CC,
-    FibreChannel = 0x8906,
-    /// RDMA over Converged Ethernet (RoCE)
-    RoCE = 0x8915,
-    LoopbackIeee8023 = 0x9000,
-}
-
-impl TryFrom<u16> for EtherType {
-    type Error = ();
-    fn try_from(value: u16) -> Result<Self, Self::Error> {
-        match value {
-            0x0060 => Ok(EtherType::Loop),
-            0x0800 => Ok(EtherType::Ipv4),
-            0x0806 => Ok(EtherType::Arp),
-            0x0842 => Ok(EtherType::WakeOnLan),
-            0x2000 => Ok(EtherType::CDP),
-            0x22EA => Ok(EtherType::SRP),
-            0x22F0 => Ok(EtherType::AVTP),
-            0x22F3 => Ok(EtherType::TRILL),
-            0x6002 => Ok(EtherType::MOP),
-            0x6003 => Ok(EtherType::DECnet),
-            0x6004 => Ok(EtherType::DECLAT),
-            0x8035 => Ok(EtherType::RARP),
-            0x809B => Ok(EtherType::AppleTalk),
-            0x80F3 => Ok(EtherType::AARP),
-            0x8100 => Ok(EtherType::VLAN),
-            0x8102 => Ok(EtherType::SLPP),
-            0x8103 => Ok(EtherType::VLACP),
-            0x86DD => Ok(EtherType::Ipv6),
-            0x8847 => Ok(EtherType::MPLSUnicast),
-            0x8848 => Ok(EtherType::MPLSMulticast),
-            0x8809 => Ok(EtherType::LACP),
-            0x88A8 => Ok(EtherType::QinQ),
-            0x88CC => Ok(EtherType::LLDP),
-            0x8906 => Ok(EtherType::FibreChannel),
-            0x8915 => Ok(EtherType::RoCE),
-            0x9000 => Ok(EtherType::LoopbackIeee8023),
-            _ => Err(()),
-        }
-    }
-}
-
-impl Default  for EtherType{
-    fn default() -> Self {
-        EtherType::Ipv4
-    }
-}
-
-impl EtherType{
-    pub fn is_vlan(&self) -> bool{
-        self == &EtherType::VLAN || self == &EtherType::QinQ
-    }
-}
